server/requirements.txt
django
pytube
django-rest-framework
requests
django-sequences
django-cors-headers
celery[redis]
drf-yasg
gunicorn
dj-static
youtube-dl
yt-dlp
